// frontend/src/generated/api/models/FarmInfoDetail.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FarmInfoDetail = {
    farmName: string;
    farmYear: number;
    farmSeasons: Array<number>;
    farmLength: string;
    farmWidth: string;
    furrowNumber?: number;
    furrowWidth?: string;
};